Third Update

+++ src/il/ac/hit/couponsproject/model/dto/ShoppingCart.java
@@ -3,8 +3,11 @@ package il.ac.hit.couponsproject.model.dto;
 import il.ac.hit.couponsproject.exception.CouponException;
 import il.ac.hit.couponsproject.model.dao.ICouponsDAO;
 import il.ac.hit.couponsproject.model.dao.impl.HibernateCouponsDAO;
-
 import java.util.*;
+import java.util.Date;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
 
 public class ShoppingCart
 {
@@ -48,6 +51,8 @@ public class ShoppingCart
 	
 	public String getXMLTable()
 	{
+		DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
+		Date date = new Date();
 		StringBuffer sb = new StringBuffer();
 		Coupon currentCoupon = null;
 		sb.append("<table class='table table-bordered table-striped' border='fixed'><thead><tr><th>תמונה</th><th>קוד קופון</th><th>שם קופון</th><th>תיאור</th><th>מחיר לפני הנחה</th><th>אחוז הנחה</th><th>מחיר חדש</th><th>תאריך תפוגה</th><th>מיקום</th><th>קטגוריה</th><th>הסר מהעגלה</th></tr></thead><tbody>");
@@ -56,19 +61,23 @@ public class ShoppingCart
 		{
 			ShoppingCartLine line = (ShoppingCartLine)iterator.next();
 			currentCoupon = line.getCoupon();
-			sb.append("<td><img src="+currentCoupon.getImage()+" height=50 width=50 /></td>");
-			sb.append("<td>"+currentCoupon.getId()+"</td>");
-			sb.append("<td>"+currentCoupon.getName()+"</td>");
-			sb.append("<td>"+currentCoupon.getDescription()+"</td>");
-			sb.append("<td>"+currentCoupon.getPrice()+"</td>");
-			sb.append("<td>"+currentCoupon.getDiscount()+"</td>");
-			sb.append("<td>"+currentCoupon.getNewprice()+"</td>");
-/*			sb.append("<td>"+currentCoupon.getLongitude()+"</td>");
-			sb.append("<td>"+currentCoupon.getLatitude()+"</td>");*/
-			sb.append("<td>"+currentCoupon.getExpiredate()+"</td>");
-			sb.append("<td>"+currentCoupon.getLocation()+"</td>");
-			sb.append("<td>"+currentCoupon.getCategory()+"</td>");
-			sb.append("<td><a href=removeFromCart?id=" + currentCoupon.getId() +" class='btn btn-primary' href='delete-coupons' role='button'>הסר</a></td></tr>");
+			System.out.println(currentCoupon.getExpiredate());
+			if(currentCoupon.getExpiredate().compareTo(dateFormat.format(date)) >0)
+			{
+				sb.append("<td><img src="+currentCoupon.getImage()+" height=50 width=50 /></td>");
+				sb.append("<td>"+currentCoupon.getId()+"</td>");
+				sb.append("<td>"+currentCoupon.getName()+"</td>");
+				sb.append("<td>"+currentCoupon.getDescription()+"</td>");
+				sb.append("<td>"+currentCoupon.getPrice()+"</td>");
+				sb.append("<td>"+currentCoupon.getDiscount()+"</td>");
+				sb.append("<td>"+currentCoupon.getNewprice()+"</td>");
+/*				sb.append("<td>"+currentCoupon.getLongitude()+"</td>");
+				sb.append("<td>"+currentCoupon.getLatitude()+"</td>");*/
+				sb.append("<td>"+currentCoupon.getExpiredate()+"</td>");
+				sb.append("<td>"+currentCoupon.getLocation()+"</td>");
+				sb.append("<td>"+currentCoupon.getCategory()+"</td>");
+				sb.append("<td><a href=removeFromCart?id=" + currentCoupon.getId() +" class='btn btn-primary' href='delete-coupons' role='button'>הסר</a></td></tr>");
+			}
 		}
 		sb.append("</tbody></table>");
 		return sb.toString();
